adding security and cache to spring profiles

+++ user/src/main/resources/sql-queries/user-protein-lists/delete-protein_list_items.sql
@@ -1,2 +1,3 @@
 DELETE FROM np_users.protein_list_items
 WHERE list_id = :list_id
+and accession_number in (:accession_numbers)
